objects/twinkle: merge twinkle control in TRX
