+++ src/libtrx/include/libtrx/game/objects/effects/twinkle.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "../../items.h"
+
+void Twinkle_SparkleItem(const ITEM *item, uint32_t mesh_mask);
