--- src/tr1/game/objects/effects/twinkle.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <libtrx/game/items.h>
-
-void Twinkle_SparkleItem(ITEM *item, uint32_t mesh_mask);
