+++ src/tr1/game/spawn.h
@@ -25,5 +25,3 @@ int16_t Spawn_GunShotHit(
 int16_t Spawn_GunShotMiss(
     int32_t x, int32_t y, int32_t z, int16_t speed, int16_t y_rot,
     int16_t room_num);
-
-void Spawn_Twinkle(const GAME_VECTOR *pos);
